Now removes all necessary fields when removing account (#105)
* Now removes all necessary fields when removing account

No longer checks first if the txo is secreted or received, as this is causing errors if the txo remains in the DB but there is not account pointing to it. Should be fine, but would like to consider removing any TXO's from the DB that no accounts point to anymore, or creating individual TXO entries for each account that points to it. Not sure what is preferred.

* Removed interaction with GiftCodes

* removing commented out code

* removed unnecessary logs

* removed print statements

* Downgrade logging of entire Tx object to trace.

* more rust fmt stuff (as usual)

Co-authored-by: Christian Oudard <christian.oudard@gmail.com>