Added Test & Changed Return
split txo is now build_split_txo_transaction and returns a tx_proposal instead of performing the entire transaction. This is in part for the sake of testing in e2e, but also to allow users to review the tx_proposal before submitting it using the submit_transaction API call.