Clean up block-index vs block-height.
A block index is the index number of a specific block. A block height is
a count of a range of blocks. This was not consistent across the API and
codebase.

Change the API for account and ledger syncing to use block heights. This
also fixes some bugs in sync reporting during offline mode.

Update the command line client to handle the new API.

Update the docs in API.md