Clean up usage of block-index vs block-height (#196)
* Add an option to `mobcli export` to show entropy onscreen.

* Clean up block-index vs block-height.

A block index is the index number of a specific block. A block height is
a count of a range of blocks. This was not consistent across the API and
codebase.

Change the API for account and ledger syncing to use block heights. This
also fixes some bugs in sync reporting during offline mode.

Update the command line client to handle the new API.

Update the docs in API.md

* GitBook: [feature/block-height] 4 pages modified

* Remove legacy docs for database usage, now that gitbook has them. Clean up gitbook errors.

* Remove legacy API.md. Remove spurious files created by gitbook.

* Update usage of block_index in gitbook docs.

* Rust formatting.

Co-authored-by: Brian <brian.william.corbin@gmail.com>