Fix a bug causing sync to create many tiny chunks.
1) Start full-service with an empty ledger and an empty wallet db
2) Import an account from entropy, giving it a first_block of 0
3) The ledger sync process will enter one block at a time into the
ledger DB. The account sync process will then attempt to decode
transactions from these blocks in very small chunks, of size 5-10
(instead of ten thousand), with no sleeping in between. Furthermore, the
first account will do all of the syncing, not letting any other accounts
sync at the same time.

In order to fix this, the condition for reporting whether the sync
thread can pause was updated to more elegantly handle this condition, by
counting the number of blocks synced, and comparing it to the ideal
chunk size.