Fix a crash when a helm chart label ends in a period.
Helm and Kubernetes require the tag for running charts to be no longer
than 63 characters, and not end in punctuation. We were handling this
accurately for truncating at a hyphen, but not at a period. The sweet
spot for triggering this is to have a branch name of length 34,
like this PR does.