Remove command substitution in help message
Previously there was a command substitution '`cargo check`' in the help
message for `tools/build-fs.sh`. This resulted in `cargo check` being
executed when the help message was printed. Now single quotes are used
around the 'cargo check' text to separate it from the rest of the text
in the help message.